Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    int num1, num2, num3, greatest;
+    float num1, num2, num3, greatest;
 
     scanf("%d %d %d", &num1, &num2, &num3);
